Workaround for broke CI with YARD 0.9.20 when using Ruby 2.7.0-dev
This PR solves the following "The manual directory is out of sync" error
when using Ruby 2.7.0-dev.

```console
% ruby -v
ruby 2.7.0dev (2019-12-03T05:51:14Z master e42d9d8df8) [x86_64-darwin17]
% CI=true bundle exec rake

(snip)

The manual directory is out of sync. Run `rake
generate_cops_documentation` and commit the results.
```

https://circleci.com/gh/rubocop-hq/rubocop-minitest/968

This issue will be resolved with the release of YARD 0.9.21 or higher,
which includes the following changes:
https://github.com/lsegal/yard/pull/1290