Make `Minitest/UnreachableAssertion` aware of `assert` and `refute` prefix methods
This PR makes `Minitest/UnreachableAssertion` aware of `assert` and `refute` prefix methods
and removes `MinitestExplorationHelpers#assertion?` method.
It makes the detection to assertion methods consistent.

NOTE: The `MinitestExplorationHelpers#assertion?` method is redundant with
`MinitestExplorationHelpers#assertion_method?` OTOH, the implementation uses
`MinitestExplorationHelpers#assertion` instead of `MinitestExplorationHelpers#assertion_method?`.
Maybe it's enough to verify that it starts with `assert` or `refute` instead of
exact match of the known assertion method names.