Added new Minitest/DuplicateTestRun cop.
If a Minitest class inherits from another class, it will also inherit its methods causing Minitest to run the parent’s tests methods twice.

This PR detects when there are two tests classes, one inherits from the other, and both have tests methods.  MinitestRubocop will add an offence to the Child class in such a case.

Note that this cop only works if both classes are on the same file.