Fix an incorrect autocorrect for `Minitest/AssertMatch`
This PR fixes an incorrect autocorrect for `Minitest/AssertMatch`
when `assert` with `match` and RHS is a regexp literal.

A regular expression literal must be the first argument to `assert_match`.
`TypeError: no implicit conversion of Regexp into String` will occur if it is
passed as the second argument.

```ruby
assert_match(object, /regexp/) #=> TypeError: no implicit conversion of Regexp into String
```

This issue was found on https://github.com/faker-ruby/faker/pull/2556.