Fix an incorrect autocorrect for `Minitest/RefuteMatch`
Follow up #181 and #184.

This PR fixes an incorrect autocorrect for `Minitest/AssertMatch`
when `refute` with `match` and RHS is a regexp literal.

A regular expression literal must be the first argument to `refute_match`.
`TypeError: no implicit conversion of Regexp into String` will occur if it is
passed as the second argument.

```ruby
refute_match(object, /regexp/) #=> TypeError: no implicit conversion of Regexp into String
```