Fix flaky testing for `Minitst/MultipleAssertions`
The shared `@cop` instance variable of `AssertOffense` module
causes flaky tests due to state changes.

The `AssertOffense` testing module needs to be updated to allow for tests
that are independent in configuration between tests.
It is necessary to prevent race conditions caused by shared configurations.

This is a workaround to skip tests until the issue is resolved in `AssertOffense` module.