Accessing an unregistered cop raises useful error
Currently, when a test class includes `AssertOffense`,
`AssertOffense#setup` attempts to dynamically identify the Cop under
test. If the Cop class cannot be found, `#setup` exits early.

In these cases, the test cases continue to execute with a nil `@cop`
since it wasn't set in `#setup` as expected. When helpers such as
`#assert_offense` are used, they access `@cop` and the tests can fail
with cryptic messages such as:

```
NoMethodError: undefined method `[]=' for nil
```

Initialization of the Cop was refactored to a lazily-called accessor
method which has the same caching functionality via `@cop`, but raises
an error if the expected Cop class isn't defined.