Add `{Assert,Refute}Match` `OnlyRegexpLiteral` config
`AssertMatch` & `RefuteMatch` enforce that

    assert matcher.match?(object)
    refute matcher.match?(object)

be replaced with

    assert_match(matcher, object)
    refute_match(matcher, object)

respectively, which use `=~` under the hood. This makes the cops highly
succeptible to false positives, in cases where the matcher object
responds to `match?`, but not to `=~`

   class SuitColorMatcher
     BLACK = [:clubs,    :spades]
     RED   = [:diamonds, :hearts]
     SUITS = BLACK + RED

     def initialize(suit)
       raise ArgumentError, suit.inspect unless SUITS.include?(suit)
       @suit = suit
     end

     def match?(other)
       case @suit
       when *BLACK then BLACK.include?(other)
       when *RED   then RED  .include?(other)
       end
     end
   end

   SuitColorMatcher.new(:clubs).match?(:spades) # => true
   SuitColorMatcher.new(:clubs) =~ :spades      # NoMethodError

or cases where the matcher object does also respond to `=~`, but in an
incompatible way

    "".match? "" # => true
    "" =~ ""     # TypeError

Given RuboCop doesn't have runtime type information, the only case we
can reliably identify is the case where a `Regexp` literal is used.

Therefore, this adds an `OnlyRegexpLiteral` config (false by default),
which consumers can enable if their codebase has many false positives.