[Fix #334] Fix minitest extension for the cop generator
Instead of reopening the Generator class in a file with the same name
of the Generator class in the main rubocop gem, we are now defining a
module that get included in the Generator class, defined in a different
file path, specific for this minitest plugin.

This allows rubocop plugins that want to use minitest to test their code
to be able to generate their own cops using the minitest extension.

Fixes #334.